+++ R/eppo_tabletools.R
@@ -1,8 +1,11 @@
 #' EPPO table manipulation tools
 #'
-#' \code{eppo_tabletools_names} creates table with names.
-#' \code{eppo_tabletools_hosts} creates table with hosts.
-#' \code{eppo_tabletools_cat} creates table with categorization.
+#' \code{eppo_tabletools_names} creates table with names -- preferred, common
+#' and synonyms -- of pests.
+#' \code{eppo_tabletools_hosts} creates table with hosts of pests.
+#' \code{eppo_tabletools_cat} creates table with categorization of pests.
+#' \code{eppo_tabletools_taxo} creates table with taxonomy of pests. All
+#' functions return both long tables and compact, human friendly tables
 #'
 #' @param names_tables A list of tables created via {\link{eppo_names_tables}}.
 #' @param token An object containing EPPO API token created via
@@ -89,34 +92,17 @@ eppo_tabletools_hosts <- function(names_tables, token) {
   hosts_table <- lapply(hosts_download, function(x) dplyr::bind_rows(x)) %>%
     bind_rows(.id = 'pest_code') %>%
     rename(host_eppocode = .data$eppocode, eppocode = .data$pest_code)
-  #use data from long table to create compact table with all host in one cell
-  #per each pest
-  nested_hosts <- hosts_table %>%
-    select(.data$labelclass, .data$full_name, .data$eppocode) %>%
-    tidyr::nest(.data$labelclass, .data$full_name)
 
-  hostIndex <- setNames(vector("list", length(eppocodes)), eppocodes)
-
-  #loop over long table to colapse all the host names into one string,
+  #take long table and colapse all the host names into one string,
   #separeted with names of host categories (major, minor, incidental etc.)
-  for (i in 1:length(nested_hosts$data)) {
-    nested_hosts$data[[i]] %>%
-      group_by(.data$labelclass) %>%
-      mutate(temp_names = paste(.data$full_name,
-                                collapse = ', ')) %>%
-      distinct(.data$temp_names) %>%
-      mutate(temp_names = paste(.data$labelclass,
-                                .data$temp_names,
-                                sep = ': ')) %>%
-      ungroup() %>%
-      select(.data$temp_names) %>%
-      transmute(hosts = paste(.data$temp_names, collapse = '; ')) %>%
-      distinct() -> hostIndex[i]
-  }
-
-  compact_table <- lapply(hostIndex,
-         function(hosts) data.frame(hosts, stringsAsFactors = FALSE)) %>%
-    bind_rows(.id = 'eppocode')
+  compact_table <- hosts_table %>%
+    group_by(.data$eppocode) %>%
+    select('labelclass', 'full_name') %>%
+    mutate(hosts = paste(.data$full_name, collapse = ', ')) %>%
+    mutate(hosts = paste(.data$labelclass, ': ', .data$hosts, collapse = '; ')) %>%
+    ungroup() %>%
+    select('eppocode', 'hosts') %>%
+    distinct()
 
   return(list(long_table = hosts_table,
                 compact_table))
@@ -132,10 +118,10 @@ eppo_tabletools_cat <- function(names_tables, token) {
   } else {
     #create reusable variables to access EPPO API
     eppocodes <- names_tables[[3]]$eppocode
-    hosts_urls <-paste0('https://data.eppo.int/api/rest/1.0/taxon/',
+    cat_urls <-paste0('https://data.eppo.int/api/rest/1.0/taxon/',
                         eppocodes, '/categorization', token)
     #download data on categorization from EPPO and strore them as list of tables
-    cat_list_table <- lapply(hosts_urls,
+    cat_list_table <- lapply(cat_urls,
                              function(x) jsonlite::fromJSON(RCurl::getURL(x)))
     cat_tables <- setNames(vector("list", length(eppocodes)), eppocodes)
     #exchange empty lists with NA tables -> NEEDS REFACTORING!!!
@@ -189,4 +175,50 @@ eppo_tabletools_cat <- function(names_tables, token) {
     return(list(list_table = cat_tables,
               compact_table))
 }
+}
+#' @rdname eppo_tabletools
+#' @export
+eppo_tabletools_taxo <- function(names_tables, token) {
+  if (!all(class(token) == c('pestr_token', 'character'))) {
+    message('Your token argument is not of pestr_token class.
+            Please provide token created with create_eppo_token function')
+  } else {
+    eppocodes <- names_tables[[3]]$eppocode
+    taxo_urls <-paste0('https://data.eppo.int/api/rest/1.0/taxon/',
+                        eppocodes, '/taxonomy', token)
+    taxo_list_table <- lapply(taxo_urls,
+                              function(x) jsonlite::fromJSON(RCurl::getURL(x)))
+    #Create empty table and NA table
+    empty_taxo_df <- data.frame(codeid = NA,
+                                eppocode = NA,
+                                prefname = NA,
+                                level = NA)
+    taxo_tables <- setNames(vector("list", length(eppocodes)), eppocodes)
+
+    #If table from list is empty exchange it with NA table
+    for (i in 1:length(taxo_list_table)) {
+      if (rlang::is_empty(taxo_list_table[[i]]) == TRUE) {
+        taxo_tables[[i]] <- empty_taxo_df
+      } else {
+        taxo_tables[i] <- taxo_list_table[i]
+      }
+    }
+    #Create compact table
+    compact_table <- data.frame(eppocode = eppocodes,
+                                taxonomy = rep(NA, length(eppocodes)),
+                                stringsAsFactors = FALSE)
+
+    for (i in 1:length(eppocodes)) {
+      if(taxo_tables[[i]][1, 3] == 'Animalia') {
+        compact_table[i, 2] <- taxo_tables[[i]][2, 3]
+      } else if (taxo_tables[[i]][1, 3] == 'Viruses and viroids'){
+        compact_table[i, 2] <- taxo_tables[[i]][2, 3]
+      } else {
+        compact_table[i, 2] <- taxo_tables[[i]][1, 3]
+      }
+    }
+
+    return(list(list_table = taxo_tables,
+                compact_table = compact_table))
+  }
 }